requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
black==26.3.1
celery==5.6.3
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==6.0.4
django-celery-beat==2.9.0
django-debug-toolbar==6.3.0
django-timezone-field==7.2.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
drf-spectacular==0.29.0
flake8==7.3.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.1
pathspec==1.0.4
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.12.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytokens==0.4.1
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
